test(issue-92): cover FAMS/FAMC back-pointers in INDI record serialization
Extract buildIndiRecord (and its PersonNode type) from scripts/export-gedcom.ts
into src/lib/gedcom.ts alongside the other pure GEDCOM helpers, and add unit
tests asserting that FAMS and FAMC back-pointers are emitted when a person is
both a spouse in one union and a child in another.
