+++ scripts/export-gedcom.ts
@@ -2,21 +2,14 @@ import * as fs from 'fs'
 import * as path from 'path'
 import neo4j from 'neo4j-driver'
 import { loadLocalEnv, validateRequiredEnv } from '../src/lib/env'
-import { GEDCOM_TYPES, escapeGedcomValue, buildNoteLines } from '../src/lib/gedcom'
+import {
+  GEDCOM_TYPES,
+  escapeGedcomValue,
+  buildIndiRecord,
+  type PersonNode,
+} from '../src/lib/gedcom'
 
 loadLocalEnv()
-
-interface PersonNode {
-  gedcomId: string
-  name: string
-  sex: string
-  birthYear: string | null
-  deathYear: string | null
-  birthPlace: string | null
-  deathPlace: string | null
-  occupation: string | null
-  notes: string | null
-}
 
 interface UnionNode {
   gedcomId: string
@@ -36,57 +29,6 @@ function groupByUnionId(rels: PersonUnionRel[]): Map<string, PersonUnionRel[]> {
     map.get(rel.unionId)!.push(rel)
   }
   return map
-}
-
-function addLifeEvent(lines: string[], tag: string, year: string | null, place: string | null): void {
-  if (!year && !place) return
-  lines.push(`1 ${tag}`)
-  if (year) lines.push(`2 ${GEDCOM_TYPES.DATE} ${year}`)
-  if (place) lines.push(`2 ${GEDCOM_TYPES.PLACE} ${escapeGedcomValue(place)}`)
-}
-
-function buildIndiRecord(person: PersonNode, famsIds: string[], famcIds: string[]): string {
-  const lines: string[] = []
-
-  lines.push(`0 ${person.gedcomId} ${GEDCOM_TYPES.INDIVIDUAL}`)
-
-  if (!person.name || person.name === '[Unknown]') {
-    lines.push(`1 ${GEDCOM_TYPES.NAME} [Unknown]`)
-  } else {
-    const parts = person.name.trim().split(' ')
-    const surname = parts.length > 1 ? parts[parts.length - 1] : ''
-    const givenName = parts.length > 1 ? parts.slice(0, -1).join(' ') : parts[0]
-    const gedcomName = surname
-      ? `${escapeGedcomValue(givenName)} /${escapeGedcomValue(surname)}/`
-      : escapeGedcomValue(givenName)
-    lines.push(`1 ${GEDCOM_TYPES.NAME} ${gedcomName}`)
-    if (givenName) lines.push(`2 ${GEDCOM_TYPES.GIVEN_NAME} ${escapeGedcomValue(givenName)}`)
-    if (surname) lines.push(`2 ${GEDCOM_TYPES.SURNAME} ${escapeGedcomValue(surname)}`)
-  }
-
-  if (person.sex) {
-    lines.push(`1 ${GEDCOM_TYPES.SEX} ${person.sex}`)
-  }
-
-  addLifeEvent(lines, GEDCOM_TYPES.BIRTH, person.birthYear, person.birthPlace)
-  addLifeEvent(lines, GEDCOM_TYPES.DEATH, person.deathYear, person.deathPlace)
-
-  if (person.occupation) {
-    lines.push(`1 ${GEDCOM_TYPES.OCCUPATION} ${escapeGedcomValue(person.occupation)}`)
-  }
-
-  if (person.notes) {
-    lines.push(...buildNoteLines(1, person.notes))
-  }
-
-  for (const uid of famsIds) {
-    lines.push(`1 ${GEDCOM_TYPES.FAMS} ${uid}`)
-  }
-  for (const uid of famcIds) {
-    lines.push(`1 ${GEDCOM_TYPES.FAMC} ${uid}`)
-  }
-
-  return lines.join('\n')
 }
 
 interface FamilyBuildContext {
